+++ parser/config.go
@@ -134,7 +134,9 @@ var SupportedFunctions = map[string]int{
 	"ispal":     1, // palindrome check
 	"fib":       1, // Fibonacci number
 	"powmod":    3, // modular exponentiation a^b mod m
-	"collatz":   1, // Collatz stopping time
+	"collatz":   1,
+	"totient":   1,
+	"sigma":     1, // Collatz stopping time
 	"prod":      -1,
 	"min":       -1, // variadic
 	"max":       -1, // variadic